feat: serve docs via GitHub Pages
This uses setup pioneered in near-api-js:
https://github.com/near/near-api-js/pull/289;
The description below is mostly copied from there.

---

Previously, we were copying the output of `yarn doc` from this
repository and commiting it to the `docs` folder in
https://github.com/nearprotocol/docs. Since `nearprotocol/docs` is a
[docusaurus](https://docusaurus.io/) website, we used
`typedoc-plugin-markdown` here in this repository to build documentation
in the format expected by docusaurus.

Recently, realizing that this `yarn doc` output is an auto-generated
artifact that we don't need to commit into either repository, I looked
into automating this whole "build docs, copy to other repository"
process; see https://github.com/nearprotocol/docs/pull/291. While
attempting to finalize this approach with @amgando, we realized it had
multiple issues:

* After about three days of effort, my script still had some major gaps
  that left the user experience lacking:

  * The sidebar on docs.nearprotocol.com did not work on most API
    Reference pages
  * "Edit this page" links worked, but not well

* Even if we spent time fixing those issues, we weren't going to be able
  to make a user experience for docs.nearprotocol.com that beat linking
  to an external docs website. Hosting docs at a consistent URL just
  doesn't give a big enough UX win to justify the maintenance burden
  this approach involves.

This last point ultimately made us scrap the approach. Why not keep
high-level info on docs.nearprotocol.com, and put API Reference docs
elsewhere? This allows each repository to maintain its own API Reference
documentation, using a design aesthetic familiar to the language
community of that repository. This seems like both a UX improvement and
a maintainability improvement for `nearprotocol/docs`.

So, this commit adds the machinery we need to automatically build and
deploy docs on every push to `master` here in `near-sdk-as`:

* remove stale `apidoc` directory from git & gitignore (missed in
  bbbf08a352332c7c03e0c9de5e84740e529f2024)
* remove `typedoc-plugin-markdown` package, as docs are no longer being
  exported to `nearprotocol/docs` repo and do not need to be built in
  docusaurus format
* simplify configuration for `typedoc` command (`--out docs` is default)
* add github workflow for deploying docs on every push to `master`
* in addition to running `typedoc`, add `docs/.nojekyll` file to tell
  GitHub that this is not a Jekyll site, otherwise all paths that start
  with an underscore throw 404 errors