Adding a new panic method that display custom Error
There is no official or legit way to create panic when certain criteria are not matched in the near assembly runtime library. Throwing a new Error `throw new Error("...")` doesn't help to debug the code. So we can add a custom panic error method that can be invoked with a custom message...
`
function panic_str(msg: string): void {
   env.panic_utf8(msg.length, changetype<usize>(String.UTF8.encode(msg, false)))
}